requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.12.1

# Data & ML
numpy==1.24.3
pandas==2.2.2
scikit-learn==1.3.0
joblib==1.3.2

# Trading & Finance
python-binance==1.0.19
yfinance==0.2.18
ccxt==4.0.75

# Utilities
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
